Fixed error with nLayout
nLayout now is long in all calls and functions, there should be no error.
Also added autoplace to the second layout in the demo.
Added a bunch of comments on the configurations, still way to go to be good.
Finally there is an issue with creating new demo, the active cells goes all the way to the last column. Need to fix that.